+++ Dominion/src/Dominion/Renderer/Buffer.h
@@ -10,7 +10,7 @@ namespace Dominion {
 	class VertexBuffer : public Bindable
 	{
 	public:
-		static Ref<VertexBuffer> Create(const float* vertices, size_t size);
+		static Ref<VertexBuffer> Create(const float* vertices, uint32_t size);
 	};
 
 	class IndexBuffer : public Bindable
@@ -18,7 +18,7 @@ namespace Dominion {
 	public:
 		virtual unsigned int GetCount() const = 0;
 
-		static Ref<IndexBuffer> Create(const uint32_t* indices, unsigned int count);
+		static Ref<IndexBuffer> Create(const uint32_t* indices, uint32_t count);
 	};
 
 }
